deleting unused controller files
