+++ src/resource-sharing/config/test/ut_template.js
